:truck: Implement app specific landing page for micro apps
