+++ libs/features/micro-app/src/lib/pages/micro-app-start-page/micro-app-start-page.component.scss
@@ -2,152 +2,8 @@
   box-sizing: border-box;
   overflow-x: hidden !important;
 }
-.main-content{
-  padding: 5% 20%;
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  gap: 4rem;
-  overflow-x: hidden !important;
-  .img-section{
-    width: 60%;
-    height: 80%;
-  }
-
-  .text-section{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-    font-size: 20px;
-    font-style: normal;
-    font-family: 'Roboto';
-    padding-top: 5%;
-    .p-text{
-      text-align: center;
-    }
-    .teal-text{
-      color: var(--micro-apps-teal-primary);
-    }
-
-    .btn-section{
-      margin-top: 22%;
-      cursor: pointer;
-      .teal-btn{
-        cursor: pointer;
-        background-color: var(--micro-apps-teal-primary);
-        width: 200px;
-        height: 48px;
-        border: none;
-        outline: none;
-        border-radius: 4px;
-        color: var(--micro-apps-white-button-text);
-
-        &:disabled{
-          background-color: gray;
-          pointer-events: none;
-        }
-      }
-
-    }
-    
-  }
-}
-
-/** Media Querries for responsive designs */
-/* Extra Small Devices (XS) */
-@media (max-width: 575px) {
-  .main-content{
-    padding: 5% 10% 5% 15%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    flex-wrap: wrap;
-    align-items: center;
-    overflow-x: hidden !important;
-  
-    .img-section{
-      width: 100%;
-      height: 100%;
-     
-      img{
-        width: 100%;
-        height: 100%;
-      }
-    }
-  }
-}
-
-/* Small Devices (SM) */
-@media (min-width: 576px) and (max-width: 767px) {
-  .main-content{
-    padding: 5% 10% 5% 15%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    flex-wrap: wrap;
-    align-items: center;
-    overflow-x: hidden !important;
-  
-    .img-section{
-      width: 100%;
-      height: 100%;
-     
-      img{
-        width: 100%;
-        height: 100%;
-      }
-    }
-  }
-}
-
-/* Medium Devices (MD) */
-@media (min-width: 768px) and (max-width: 1199px) {
-  .main-content{
-    padding:  5% 5%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-    overflow-x: hidden !important;
-  
-    .img-section{
-      width: 100%;
-      height: 100%;
-     
-      img{
-        width: 90%;
-        height: 100%;
-      }
-    }
-  }  
-  .text-section{
-    width: 90%;
-    display: flex;
-    flex-direction: column;
-    justify-content: end;
-    align-items: end;
-  }
-}
-
- /* Extra Large Devices (XL) */
-@media (min-width: 1200px) and (max-width: 1599px) {
-  .main-content{
-    padding: 5% 20%;
-    display: grid;
-    grid-template-columns: 1fr 2fr;
-    gap: 4rem;
-    overflow-x: hidden !important;
-  }
-}
-
-/* Extra Extra Large Devices (XXL) */
-@media (min-width: 1600px) {
-  .main-content{
-    padding: 5% 20%;
-    display: grid;
-    grid-template-columns: 1fr 2fr;
-    gap: 4rem;  
-    overflow-x: hidden !important;
-  }
+.mat-spinner-div{
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
